Finish cleaning up routers
