+++ src/panels/config/integrations/integration-panels/ozw/ozw-node-dashboard.ts
@@ -25,6 +25,7 @@ import {
   OZWDeviceMetaDataResponse,
 } from "../../../../../data/ozw";
 import { showOZWRefreshNodeDialog } from "./show-dialog-ozw-refresh-node";
+import { ozwNetworkTabs } from "./ozw-network-router";
 
 @customElement("ozw-node-dashboard")
 class OZWNodeDashboard extends LitElement {
@@ -58,71 +59,78 @@ class OZWNodeDashboard extends LitElement {
 
   protected render(): TemplateResult {
     return html`
-      <ha-config-section .narrow=${this.narrow} .isWide=${this.isWide}>
-        <div slot="header">
-          Node Management
-        </div>
+      <hass-tabs-subpage
+        .hass=${this.hass}
+        .narrow=${this.narrow}
+        .route=${this.route}
+        .tabs=${ozwNetworkTabs(this.ozwInstance)}
+      >
+        <ha-config-section .narrow=${this.narrow} .isWide=${this.isWide}>
+          <div slot="header">
+            Node Management
+          </div>
 
-        <div slot="introduction">
-          View the status of a node and manage its configuration.
-        </div>
+          <div slot="introduction">
+            View the status of a node and manage its configuration.
+          </div>
 
-        ${this._node
-          ? html`
-              <ha-card class="content">
-                <div class="card-content">
-                  <b
-                    >${this._node.node_manufacturer_name}
-                    ${this._node.node_product_name}</b
-                  ><br />
-                  Node ID: ${this._node.node_id}<br />
-                  Query Stage: ${this._node.node_query_stage}
-                  ${this._metadata?.metadata.ProductManualURL
-                    ? html` <a
-                        href="${this._metadata.metadata.ProductManualURL}"
-                      >
-                        <p>Product Manual</p>
-                      </a>`
-                    : ``}
-                </div>
-                <div class="card-actions">
-                  <mwc-button @click=${this._refreshNodeClicked}>
-                    Refresh Node
-                  </mwc-button>
-                </div>
-              </ha-card>
-            `
-          : ``}
-        ${this._metadata
-          ? html`
-              <ha-card class="content" header="Description">
-                <div class="card-content">
-                  ${this._metadata.metadata.Description}
-                </div>
-              </ha-card>
-              <ha-card class="content" header="Inclusion">
-                <div class="card-content">
-                  ${this._metadata.metadata.InclusionHelp}
-                </div>
-              </ha-card>
-              <ha-card class="content" header="Exclusion">
-                <div class="card-content">
-                  ${this._metadata.metadata.ExclusionHelp}
-                </div>
-              </ha-card>
-              <ha-card class="content" header="Reset">
-                <div class="card-content">
-                  ${this._metadata.metadata.ResetHelp}
-                </div>
-              </ha-card>
-              <ha-card class="content" header="WakeUp">
-                <div class="card-content">
-                  ${this._metadata.metadata.WakeupHelp}
-                </div>
-              </ha-card>
-            `
-          : ``}
-      </ha-config-section>
+          ${this._node
+            ? html`
+                <ha-card class="content">
+                  <div class="card-content">
+                    <b
+                      >${this._node.node_manufacturer_name}
+                      ${this._node.node_product_name}</b
+                    ><br />
+                    Node ID: ${this._node.node_id}<br />
+                    Query Stage: ${this._node.node_query_stage}
+                    ${this._metadata?.metadata.ProductManualURL
+                      ? html` <a
+                          href="${this._metadata.metadata.ProductManualURL}"
+                        >
+                          <p>Product Manual</p>
+                        </a>`
+                      : ``}
+                  </div>
+                  <div class="card-actions">
+                    <mwc-button @click=${this._refreshNodeClicked}>
+                      Refresh Node
+                    </mwc-button>
+                  </div>
+                </ha-card>
+              `
+            : ``}
+          ${this._metadata
+            ? html`
+                <ha-card class="content" header="Description">
+                  <div class="card-content">
+                    ${this._metadata.metadata.Description}
+                  </div>
+                </ha-card>
+                <ha-card class="content" header="Inclusion">
+                  <div class="card-content">
+                    ${this._metadata.metadata.InclusionHelp}
+                  </div>
+                </ha-card>
+                <ha-card class="content" header="Exclusion">
+                  <div class="card-content">
+                    ${this._metadata.metadata.ExclusionHelp}
+                  </div>
+                </ha-card>
+                <ha-card class="content" header="Reset">
+                  <div class="card-content">
+                    ${this._metadata.metadata.ResetHelp}
+                  </div>
+                </ha-card>
+                <ha-card class="content" header="WakeUp">
+                  <div class="card-content">
+                    ${this._metadata.metadata.WakeupHelp}
+                  </div>
+                </ha-card>
+              `
+            : ``}
+        </ha-config-section>
+      </hass-tabs-subpage>
     `;
   }
